Move changes to react-next
